refactor: 변경된(webp) 배너 이미지로 변경
